Sending alarm status and triggers from microchip works
I've finally gotten the read in properly working from the Android side.
I also got some working code on the Android end to read in keywords and
take the rest. So statusOn will now trigger status textview to show On.
On the microchip end, I've attached a 3 1.5V pack battery  so now it's
portable.