Saving some progress made
Decided to just force D/C to change name because there's not much of an
option. The user would just have to reconnect again for the name to
change.
Other option was to have a table saved onto the phone that keeps track
of the ID = name, but I rather have it saved onto the bluetooth so the
sensor is transferable to other phones. Both ways require the user to
spend time setting up the system. I think overtime changing the
bluetooth overall is better suited for now.